S12P11E208-89 [BE] fix : 데이터 타입 변경

+++ backend/src/main/java/com/dasima/drawrun/domain/course/dto/request/CourseResultSaveRequest.java
@@ -9,10 +9,10 @@ public class CourseResultSaveRequest {
   private Integer pathId;
   private String startImgUrl;
   private String endImgUrl;
-  private double distance_km;
+  private Float distance_km;
   private Long time_s;
   private Long pace_s;
-  private Integer calorie;
+  private Float calorie;
   private Integer state;
   private Integer heartbeat;
   private Integer cadence;
